hf_space/requirements.txt
gradio>=4.44.0
ultralytics>=8.3.0
torch>=2.2.0
torchvision>=0.17.0
pillow>=10.2.0
numpy>=1.26.0
opencv-python-headless>=4.10.0.84
mmengine>=0.10.5
mmcv-lite>=2.1.0,<2.2.0
mmdet>=3.3.0
